Safari-609.1.11.1/LayoutTests/http/tests/security/cross-frame-access-protocol-explicit-domain-expected.txt

CONSOLE MESSAGE: line 194: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
This test currently fails because we check the port and protocol even if document.domain is explicitly set (rdar://problem/5366437).


PASS: Cross frame access to https from http, after explicitly setting document.domain, was denied.


--------
Frame: 'aFrame'
--------
Inner iframe with explicit document.domain set.
